Moves Ensemble and Barostat move code to attic
* No longer used and heavily relied on simtk.unit
